Factor out constants and some helper methods to new classes
